+++ exhibitera/apps/media_player/media_player.js
@@ -1,6 +1,7 @@
 /* global Js3dModelViewer */
-import * as exCommon from '../js/exhibitera_app_common.js'
-import * as exMarkdown from '../js/exhibitera_app_markdown.js'
+import * as exFiles from '../../common/files.js'
+import * as appsCommon from '../js/exhibitera_app_common.js'
+import * as appsMarkdown from '../js/exhibitera_app_markdown.js'
 
 function updateParser (update) {
   // Read updates specific to the media player
@@ -16,11 +17,59 @@ function loadDefinition (def) {
 
   const root = document.querySelector(':root')
 
-  exCommon.config.definition = def
-  exCommon.config.sourceList = []
+  appsCommon.config.definition = def
+  appsCommon.config.sourceList = []
   def.content_order.forEach((uuid) => {
-    exCommon.config.sourceList.push(def.content[uuid])
+    appsCommon.config.sourceList.push(def.content[uuid])
   })
+
+  // Create tne progress indicators
+  const progressContainer = document.getElementById('progressContainer')
+  const progressIndicator = document.getElementById('progressIndicator')
+  progressIndicator.innerText = ''
+
+  for (let i = 0; i < def.content_order.length; i++) {
+    const dot = document.createElement('div')
+    dot.classList = 'progress-dot'
+    progressIndicator.appendChild(dot)
+  }
+
+  // Indicator position
+  const posSplit = (def?.behavior?.progress_indicator?.position ?? 'bottom_left').split('_')
+  const vertPos = posSplit[0]
+  const horizPos = posSplit[1]
+
+  if (vertPos === 'top') {
+    progressContainer.style.top = 0
+  } else {
+    progressContainer.style.bottom = 0
+  }
+  if (horizPos === 'left') {
+    progressContainer.classList.add('justify-content-start')
+    progressContainer.classList.remove('justify-content-center')
+    progressContainer.classList.remove('justify-content-end')
+  } else if (horizPos === 'middle') {
+    progressContainer.classList.add('justify-content-center')
+    progressContainer.classList.remove('justify-content-start')
+    progressContainer.classList.remove('justify-content-end')
+  } else if (horizPos === 'right') {
+    progressContainer.classList.add('justify-content-end')
+    progressContainer.classList.remove('justify-content-start')
+    progressContainer.classList.remove('justify-content-center')
+  }
+
+  // Indicator size
+  const indicatorSize = parseFloat(def?.behavior?.progress_indicator?.size ?? 1)
+  root.style.setProperty('--progress-indicator-size', indicatorSize)
+  progressIndicator.style.width = String(2 * indicatorSize * (def.content_order.length - 1) + 2 * indicatorSize) + 'vw'
+
+  // Indicator visibility
+  console.log(def?.behavior?.progress_indicator?.visible)
+  if (def?.behavior?.progress_indicator?.visible ?? false) {
+    progressContainer.style.display = 'flex'
+  } else {
+    progressContainer.style.display = 'none'
+  }
 
   // Watermark settings
   const watermarkEl = document.getElementById('watermark')
@@ -40,14 +89,17 @@ function loadDefinition (def) {
 
   // Color
   // First, reset to defaults (in case a style option doesn't exist in the definition)
-  root.style.setProperty('--subtitle-color', 'white')
+  root.style.setProperty('--subtitleColor', 'white')
+  root.style.setProperty('--progressBackgroundColor', '#64646499')
+  root.style.setProperty('--progressActiveColor', '#008000')
+  root.style.setProperty('--progressInactiveColor', '#808080')
 
   // Then, apply the definition settings
   Object.keys(def?.style?.color ?? []).forEach((key) => {
     console.log(def.style.color[key])
     document.documentElement.style.setProperty('--' + key, def.style.color[key])
   })
-  exCommon.setBackground(def?.style?.background ?? {}, root, '#000', true)
+  appsCommon.setBackground(def?.style?.background ?? {}, root, '#000', true)
 
   // Font
 
@@ -80,13 +132,26 @@ function gotoSource (index) {
   // Make sure the index is an integer
   index = parseInt(index)
 
-  if (exCommon.config.debug) {
+  if (appsCommon.config.debug) {
     console.log('gotoSource', index)
   }
 
-  if (index < exCommon.config.sourceList.length) {
-    exCommon.config.activeIndex = index
-    changeMedia(exCommon.config.sourceList[index])
+  if (index < appsCommon.config.sourceList.length) {
+    appsCommon.config.activeIndex = index
+
+    // Update the progress indicator to show the current progress
+    const dots = document.querySelectorAll('.progress-dot')
+    if (dots.length === 0) return
+    for (const dot of dots) {
+      dot.classList.remove('active')
+      dot.innerText = ''
+    }
+    dots[index].classList.add('active')
+    const fillEl = document.createElement('div')
+    fillEl.classList = 'fill'
+    dots[index].appendChild(fillEl)
+
+    changeMedia(index)
   }
 }
 
@@ -94,20 +159,22 @@ function gotoNextSource () {
   // Display the next file in sourceList, looping to the beginning if
   // necessary
 
-  if (exCommon.config.activeIndex + 1 >= exCommon.config.sourceList.length) {
-    exCommon.config.activeIndex = 0
+  if (appsCommon.config.activeIndex + 1 >= appsCommon.config.sourceList.length) {
+    appsCommon.config.activeIndex = 0
   } else {
-    exCommon.config.activeIndex += 1
+    appsCommon.config.activeIndex += 1
   }
 
-  gotoSource(exCommon.config.activeIndex)
+  gotoSource(appsCommon.config.activeIndex)
 }
 
-async function changeMedia (source) {
+async function changeMedia (index) {
   // Load and play a media file given in source
   // delayPlay and playOnly are used when synchronizing multiple displays
 
-  if (exCommon.config.debug) {
+  const source = appsCommon.config.sourceList[index]
+
+  if (appsCommon.config.debug) {
     console.log('changeMedia', source)
   }
 
@@ -121,7 +188,7 @@ async function changeMedia (source) {
 
   let filename
   if (('type' in source) === false || source.type === 'file') {
-    filename = exCommon.config.helperAddress + '/content/' + source.filename
+    filename = appsCommon.config.helperAddress + '/content/' + source.filename
   } else if (source.type === 'url') {
     filename = source.filename
   }
@@ -130,11 +197,16 @@ async function changeMedia (source) {
     filename += (/\?/.test(filename) ? '&' : '?') + new Date().getTime()
   }
 
+  const mimetype = exFiles.guessMimetype(source.filename)
+  if (['audio', 'image', 'model', 'video'].includes(mimetype) === false) {
+    console.log(`File ${source.filename} has an invalid mimetype of ${mimetype}`)
+  }
+
   // Split off the extension
   const split = source.filename.split('.')
   const ext = split[split.length - 1].toLowerCase()
 
-  if (['mp4', 'mpeg', 'm4v', 'webm', 'mov', 'ogv', 'mpg'].includes(ext)) {
+  if (mimetype === 'video') {
     // Video file
     clearTimeout(sourceAdvanceTimer) // Only used for pictures
     videoContainer.style.opacity = 1
@@ -145,20 +217,21 @@ async function changeMedia (source) {
       video.pause()
       video.src = filename
       video.load()
-      video.play()
     }
+    video.play()
+
     // Subtitles
     if (subtitleEl != null) subtitleEl.remove()
     if (source?.subtitles?.filename != null) {
       subtitleEl = document.createElement('track')
-      subtitleEl.src = exCommon.config.helperAddress + '/content/' + source.subtitles.filename
+      subtitleEl.src = appsCommon.config.helperAddress + '/content/' + source.subtitles.filename
       video.appendChild(subtitleEl)
       video.textTracks[0].mode = 'showing'
     }
-    if (exCommon.config.sourceList.length > 1) { // Don't loop or onended will never fire
+    if (appsCommon.config.sourceList.length > 1) { // Don't loop or onended will never fire
       video.loop = false
       video.onended = function () {
-        if (exCommon.config.autoplayEnabled === true) {
+        if (appsCommon.config.autoplayEnabled === true) {
           gotoNextSource()
         } else {
           video.play()
@@ -167,7 +240,7 @@ async function changeMedia (source) {
     } else {
       video.loop = true
     }
-  } else if (['png', 'jpg', 'jpeg', 'tiff', 'tif', 'bmp', 'heic', 'webp'].includes(ext)) {
+  } else if (mimetype === 'image') {
     // Image file
     video.pause()
     audio.pause()
@@ -177,7 +250,8 @@ async function changeMedia (source) {
     imageContainer.style.opacity = 1
     clearTimeout(sourceAdvanceTimer)
     sourceAdvanceTimer = setTimeout(gotoNextSource, source.duration * 1000)
-  } else if (['aac', 'm4a', 'mp3', 'oga', 'ogg', 'weba', 'wav'].includes(ext)) {
+    trackProgress(index, source.duration * 1000)
+  } else if (mimetype === 'audio') {
     // Audio file
     video.pause()
     videoContainer.style.opacity = 0
@@ -188,12 +262,13 @@ async function changeMedia (source) {
       audio.pause()
       audio.src = filename
       audio.load()
-      audio.play()
     }
-    if (exCommon.config.sourceList.length > 1) { // Don't loop or onended will never fire
+    audio.play()
+
+    if (appsCommon.config.sourceList.length > 1) { // Don't loop or onended will never fire
       audio.loop = false
       audio.onended = function () {
-        if (exCommon.config.autoplayEnabled === true) {
+        if (appsCommon.config.autoplayEnabled === true) {
           gotoNextSource()
         } else {
           audio.play()
@@ -202,15 +277,15 @@ async function changeMedia (source) {
     } else {
       audio.loop = true
     }
-  } else if (['obj', 'glb'].includes(ext)) {
+  } else if (mimetype === 'model') {
     // 3D model
     videoContainer.style.opacity = 0
     imageContainer.style.opacity = 0
     modelContainer.style.opacity = 1
 
     let backgroundColor = 'rgb(0, 0, 0)'
-    if (exCommon.config.definition.style.background.mode === 'color') {
-      backgroundColor = exCommon.config.definition.style.background.color
+    if (appsCommon.config.definition.style.background.mode === 'color') {
+      backgroundColor = appsCommon.config.definition.style.background.color
     }
 
     const scene = Js3dModelViewer.prepareScene(model, {
@@ -249,6 +324,7 @@ async function changeMedia (source) {
         document.getElementById('modelViewer').innerHTML = ''
       }, 500)
     }, source.duration * 1000)
+    trackProgress(index, source.duration * 1000)
   }
 
   // Annotations
@@ -266,12 +342,24 @@ async function changeMedia (source) {
   }
 }
 
+function trackProgress (index, total, elapsed = 0) {
+  // Track the progress of an image or model and update the progress dot
+
+  clearTimeout(progressTimer)
+
+  const activeDot = document.querySelector('.progress-dot.active .fill')
+  const pct = (elapsed / total) * 100
+  activeDot.style.width = `${pct}%`
+
+  progressTimer = setTimeout(() => trackProgress(index, total, elapsed + 100), 100)
+}
+
 function createAnnotation (details) {
   // Render an annotation on the display.
 
   const annotation = document.createElement('div')
   annotation.classList = 'annotation'
-  annotation.innerHTML = exMarkdown.formatText(String(details.value), { removeParagraph: true, string: true })
+  annotation.innerHTML = appsMarkdown.formatText(String(details.value), { removeParagraph: true, string: true })
   annotation.style.position = 'absolute'
 
   const xPos = details?.x_position ?? 50
@@ -286,7 +374,7 @@ function createAnnotation (details) {
   }
 
   if ('font' in details) {
-    annotation.style.fontFamily = exCommon.createFont(details.font, details.font)
+    annotation.style.fontFamily = appsCommon.createFont(details.font, details.font)
   } else {
     annotation.style.fontFamily = 'annotation-default'
   }
@@ -309,7 +397,7 @@ function fetchAnnotation (details) {
 
   return new Promise((resolve, reject) => {
     if (details.type === 'file') {
-      exCommon.makeHelperRequest({
+      appsCommon.makeHelperRequest({
         method: 'GET',
         api: '',
         endpoint: '/content/' + details.file,
@@ -345,12 +433,12 @@ function fetchAnnotation (details) {
   })
 }
 
-exCommon.config.activeIndex = 0 // Index of the file from the source list currently showing
-exCommon.config.sourceList = []
-exCommon.config.autoplayEnabled = true
+appsCommon.config.activeIndex = 0 // Index of the file from the source list currently showing
+appsCommon.config.sourceList = []
+appsCommon.config.autoplayEnabled = true
 let subtitleEl = null
 
-exCommon.configureApp({
+appsCommon.configureApp({
   name: 'media_player',
   debug: true,
   loadDefinition,
@@ -359,3 +447,20 @@ exCommon.configureApp({
 
 const currentDefintion = ''
 let sourceAdvanceTimer = null // Will hold reference to a setTimeout instance to move to the next media.
+let progressTimer = null // Will track the progress of a source towards completion
+
+// Bind event listeners
+const video = document.getElementById('fullscreenVideo')
+video.addEventListener('timeupdate', () => {
+  clearTimeout(progressTimer)
+  const activeDot = document.querySelector('.progress-dot.active .fill')
+  const pct = (video.currentTime / video.duration) * 100
+  activeDot.style.width = `${pct}%`
+})
+const audio = document.getElementById('audioPlayer')
+audio.addEventListener('timeupdate', () => {
+  clearTimeout(progressTimer)
+  const activeDot = document.querySelector('.progress-dot.active .fill')
+  const pct = (audio.currentTime / audio.duration) * 100
+  activeDot.style.width = `${pct}%`
+})